requirements.txt
alembic==1.0.11
APScheduler==3.6.1
Click==7.0
Flask==1.1.1
Flask-APScheduler==1.11.0
Flask-Cors==3.0.8
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
psycopg2==2.8.3
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
six==1.12.0
SQLAlchemy==1.3.7
tzlocal==2.0.0
Werkzeug==0.15.5
